% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-truth-data.R
\name{download_truth_data}
\alias{download_truth_data}
\title{Load Truth Data}
\usage{
download_truth_data(
  countries = c("Germany", "Poland"),
  save_dir = "data",
  create_dir = TRUE
)
}
\arguments{
\item{countries}{character vector with the country names to download}

\item{save_dir}{character, directory in which to save the data}

\item{create_dir}{logical, whether or not to create the directory if it
doesn't exist yet. Default is TRUE}
}
\value{
A data.frame
}
\description{
MISSING
}
