% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-truth-data.R
\name{load_truth_data}
\alias{load_truth_data}
\title{Load Truth Data}
\usage{
load_truth_data(
  countries = c("Germany", "Poland"),
  root_dir = "data",
  load_from_server = FALSE,
  cases = TRUE,
  weekly = TRUE,
  national_only = TRUE,
  cumulative = FALSE
)
}
\arguments{
\item{countries}{character vector with the country names to load}

\item{root_dir}{charcter. Where to look for the data to load or where to
download it to if \code{load_from_server = TRUE}}

\item{load_from_server}{logical, whether or not to download the data from
the server first. Default is FALSE}

\item{cases}{logical, whether or not to create return cases (the default)
or deaths (if \code{cases = FALSE})}

\item{weekly}{logical, whether or not to return weekly data or daily data
(if \code{weekly = FALSE})}

\item{national_only}{logical, whether or not to only return national
estimates (TRUE) or all estimates (if \code{national_only = FALSE})}

\item{cumulative}{logical, whether or not to return incident data (the
default) or cumulative data (if cumulative = TRUE)}
}
\value{
A data.frame
}
\description{
MISSING
}
