% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_weekly}
\alias{make_weekly}
\title{Make Incidence data weekly}
\usage{
make_weekly(
  data,
  value_cols = "value",
  weekstart = 7L,
  group_by = "location",
  exclude_incomplete = TRUE
)
}
\arguments{
\item{data}{A data frame containing a column called 'date'}

\item{value_cols}{character vector with the names of columns for which a
weekly value shall be computed}

\item{weekstart}{integer that indicates the start of the week. Default is 7 (Sunday)}

\item{group_by}{A character vector with variables to group by}

\item{exclude_incomplete}{exclude incomplete epiweeks (not equal to 7
observations for that week). Default is TRUE. If an empty data.table is returned,
make sure that you got the grouping in `group_by` right.}
}
\value{
A data frame grouped by week.
}
\description{
Make Incidence data weekly
}
\examples{
df <- data.frame(value = 1:30,
                 date = seq.Date(from = as.Date("2020-12-20"),
                 to = as.Date("2021-01-18"), by = "days"),
                 location = "GM",
                 location_name = "Germany")

computed <- make_weekly(df, group_by = "location")
}
